requirements.txt
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.3
customtkinter==5.2.0
darkdetect==0.8.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pyserial==3.5
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
urllib3==2.2.1
Werkzeug==3.0.1
xmltodict==0.13.0
